+++ blue-core/src/main/java/blue/soundObject/jmask/Oscillator.java
@@ -262,11 +262,6 @@ public class Oscillator implements Generator, Maskable,
     }
 
     @Override
-    public boolean equals(Object obj) {
-        return EqualsBuilder.reflectionEquals(this, obj);
-    }
-
-    @Override
     public String toString() {
         return ToStringBuilder.reflectionToString(this);
     }